requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
Pillow>=9.5.0
cryptography>=41.0.0
requests>=2.31.0
vaderSentiment>=3.3.2
textblob>=0.17.1
joblib>=1.3.0